Improve code structure and add method comment
